Fixed bug with scrolling/key press to move to frames
I noticed that when hovering over the canvas and scrolling, the current frame number incremented/decremented beyond the range of frames (for me, below 0 and above the total number of frames). The same bug happened with key left/right presses.

I added an if-statement in annotationweb.js that seems to fix the problem.