When exporting annotations on MHD images, retrieve the original aspect ratio to re-scale the annotations location to the original image spacing.
Fixed misalignment between the annotation and the segmentation when exporting anisotropic data. This will happen whenever exporting a MHD image, with offset information in the metadata.

Exposed the attribute Offset from the MHD header. This can be used to correctly align the annotations to the frame when exporting the data.